+++ crates/ecs/src/storage.rs
@@ -1,19 +1,45 @@
-use std::{
-    any::{Any, TypeId},
-    cmp::Ordering,
-};
+use std::any::{Any, TypeId};
 
 use fnv::FnvHashMap;
+use pulz_schedule::resource::FromResources;
 
-use crate::{
-    archetype::ArchetypeId,
-    resource::{Res, ResourceId, Resources},
-    Entity,
-};
+use crate::{archetype::ArchetypeId, Entity};
 
-pub enum Storage<T> {
-    Dense(Vec<Vec<T>>),
-    Sparse(FnvHashMap<Entity, T>),
+pub trait Storage: Send + Sync + Any + FromResources {
+    const SPARSE: bool;
+
+    type Component;
+
+    #[inline]
+    fn component_type_id() -> TypeId {
+        TypeId::of::<Self::Component>()
+    }
+
+    fn contains(&self, entity: Entity, archetype: ArchetypeId, index: usize) -> bool;
+    fn swap_remove(
+        &mut self,
+        entity: Entity,
+        archetype: ArchetypeId,
+        index: usize,
+    ) -> Option<Self::Component>;
+    fn insert(&mut self, entity: Entity, archetype: ArchetypeId, value: Self::Component) -> usize;
+    fn replace(
+        &mut self,
+        entity: Entity,
+        archetype: ArchetypeId,
+        index: usize,
+        value: Self::Component,
+    ) -> Option<Self::Component>;
+
+    fn get(&self, entity: Entity, archetype: ArchetypeId, index: usize)
+        -> Option<&Self::Component>;
+
+    fn get_mut(
+        &mut self,
+        entity: Entity,
+        archetype: ArchetypeId,
+        index: usize,
+    ) -> Option<&mut Self::Component>;
 }
 
 pub trait AnyStorage: Send + Sync + Any {
@@ -42,6 +68,21 @@ pub trait AnyStorage: Send + Sync + Any {
     ) -> Option<usize>;
 }
 
+pub struct DenseStorage<T>(Vec<Vec<T>>);
+pub struct HashMapStorage<T>(FnvHashMap<Entity, T>);
+
+impl<T> Default for DenseStorage<T> {
+    fn default() -> Self {
+        Self(Vec::new())
+    }
+}
+
+impl<T> Default for HashMapStorage<T> {
+    fn default() -> Self {
+        Self(FnvHashMap::default())
+    }
+}
+
 fn vec_make_available<T: Default>(vec: &mut Vec<T>, index: usize) -> &mut T {
     if vec.len() <= index {
         vec.resize_with(index + 1, Default::default);
@@ -50,92 +91,146 @@ fn vec_make_available<T: Default>(vec: &mut Vec<T>, index: usize) -> &mut T {
     unsafe { vec.get_unchecked_mut(index) }
 }
 
-impl<T> Storage<T>
+impl<T> Storage for DenseStorage<T>
 where
     T: Send + Sync + 'static,
 {
-    #[inline]
-    pub fn new_dense() -> Self {
-        Self::Dense(Vec::new())
-    }
-    #[inline]
-    pub fn new_sparse() -> Self {
-        Self::Sparse(FnvHashMap::default())
-    }
-    fn take_t(value: &mut dyn Any) -> Option<T> {
-        value.downcast_mut::<Option<T>>()?.take()
-    }
-
-    pub(crate) fn from_res(res: &Resources, id: ResourceId) -> Option<Res<'_, dyn AnyStorage>> {
-        Some(Res::map(res.borrow_res_id::<Self>(id.typed())?, |s| {
-            let d: &dyn AnyStorage = s;
-            d
-        }))
-    }
-
-    pub(crate) fn from_res_mut(res: &mut Resources, id: ResourceId) -> Option<&mut dyn AnyStorage> {
-        Some(res.get_mut_id::<Self>(id.typed())?)
-    }
+    const SPARSE: bool = false;
+    type Component = T;
 
     #[inline]
-    pub fn get(&self, entity: Entity, archetype: ArchetypeId, index: usize) -> Option<&T> {
-        match self {
-            Storage::Dense(columns) => columns.get(archetype.index())?.get(index),
-            Storage::Sparse(map) => map.get(&entity),
+    fn contains(&self, _entity: Entity, archetype: ArchetypeId, index: usize) -> bool {
+        self.0
+            .get(archetype.index())
+            .map_or(false, |col| index < col.len())
+    }
+    #[inline]
+    fn swap_remove(&mut self, _entity: Entity, archetype: ArchetypeId, index: usize) -> Option<T> {
+        if let Some(col) = self.0.get_mut(archetype.index()) {
+            if index < col.len() {
+                return Some(col.swap_remove(index));
+            }
         }
+        None
     }
-
     #[inline]
-    pub fn get_mut(
+    fn insert(&mut self, _entity: Entity, archetype: ArchetypeId, value: T) -> usize {
+        let col = vec_make_available(&mut self.0, archetype.index());
+        let new_index = col.len();
+        col.push(value);
+        new_index
+    }
+    #[inline]
+    fn replace(
         &mut self,
-        entity: Entity,
+        _entity: Entity,
         archetype: ArchetypeId,
         index: usize,
-    ) -> Option<&mut T> {
-        match self {
-            Storage::Dense(columns) => columns.get_mut(archetype.index())?.get_mut(index),
-            Storage::Sparse(map) => map.get_mut(&entity),
+        value: T,
+    ) -> Option<T> {
+        let col = vec_make_available(&mut self.0, archetype.index());
+        if let Some(entry) = col.get_mut(index) {
+            Some(std::mem::replace(entry, value))
+        } else {
+            None
         }
+    }
+
+    #[inline]
+    fn get(
+        &self,
+        _entity: Entity,
+        archetype: ArchetypeId,
+        index: usize,
+    ) -> Option<&Self::Component> {
+        self.0.get(archetype.index())?.get(index)
+    }
+
+    #[inline]
+    fn get_mut(
+        &mut self,
+        _entity: Entity,
+        archetype: ArchetypeId,
+        index: usize,
+    ) -> Option<&mut Self::Component> {
+        self.0.get_mut(archetype.index())?.get_mut(index)
     }
 }
 
-impl<T> AnyStorage for Storage<T>
+impl<T> Storage for HashMapStorage<T>
 where
     T: Send + Sync + 'static,
 {
+    const SPARSE: bool = true;
+    type Component = T;
+
+    #[inline]
+    fn contains(&self, entity: Entity, _archetype: ArchetypeId, _index: usize) -> bool {
+        self.0.contains_key(&entity)
+    }
+    #[inline]
+    fn swap_remove(&mut self, entity: Entity, _archetype: ArchetypeId, _index: usize) -> Option<T> {
+        self.0.remove(&entity)
+    }
+    #[inline]
+    fn insert(&mut self, entity: Entity, _archetype: ArchetypeId, value: T) -> usize {
+        let len = self.0.len();
+        self.0.insert(entity, value);
+        len
+    }
+    #[inline]
+    fn replace(
+        &mut self,
+        entity: Entity,
+        _archetype: ArchetypeId,
+        _index: usize,
+        value: T,
+    ) -> Option<T> {
+        self.0.insert(entity, value)
+    }
+
+    #[inline]
+    fn get(
+        &self,
+        entity: Entity,
+        _archetype: ArchetypeId,
+        _index: usize,
+    ) -> Option<&Self::Component> {
+        self.0.get(&entity)
+    }
+
+    #[inline]
+    fn get_mut(
+        &mut self,
+        entity: Entity,
+        _archetype: ArchetypeId,
+        _index: usize,
+    ) -> Option<&mut Self::Component> {
+        self.0.get_mut(&entity)
+    }
+}
+
+fn take_option_t<T: 'static>(value: &mut dyn Any) -> Option<T> {
+    value.downcast_mut::<Option<T>>()?.take()
+}
+
+impl<S> AnyStorage for S
+where
+    S: Storage,
+{
     #[inline]
     fn component_type_id(&self) -> TypeId {
-        TypeId::of::<T>()
+        S::component_type_id()
     }
 
     #[inline]
     fn contains(&self, entity: Entity, archetype: ArchetypeId, index: usize) -> bool {
-        match self {
-            Storage::Dense(columns) => columns
-                .get(archetype.index())
-                .map_or(false, |col| index < col.len()),
-            Storage::Sparse(map) => map.contains_key(&entity),
-        }
+        S::contains(self, entity, archetype, index)
     }
 
     #[inline]
     fn swap_remove(&mut self, entity: Entity, archetype: ArchetypeId, index: usize) -> bool {
-        match self {
-            Storage::Dense(columns) => {
-                if let Some(col) = columns.get_mut(archetype.index()) {
-                    if index < col.len() {
-                        col.swap_remove(index);
-                        return true;
-                    }
-                }
-            }
-            Storage::Sparse(map) => {
-                if map.remove(&entity).is_some() {
-                    return true;
-                }
-            }
-        }
-        false
+        S::swap_remove(self, entity, archetype, index).is_some()
     }
 
     #[inline]
@@ -145,19 +240,8 @@ where
         archetype: ArchetypeId,
         value: &mut dyn Any,
     ) -> Option<usize> {
-        let value_t = Self::take_t(value)?;
-        match self {
-            Storage::Dense(columns) => {
-                let col = vec_make_available(columns, archetype.index());
-                let new_index = col.len();
-                col.push(value_t);
-                Some(new_index)
-            }
-            Storage::Sparse(map) => {
-                map.insert(entity, value_t);
-                Some(0)
-            }
-        }
+        let value_t = take_option_t::<S::Component>(value)?;
+        Some(S::insert(self, entity, archetype, value_t))
     }
 
     #[inline]
@@ -168,31 +252,16 @@ where
         index: usize,
         value: &mut dyn Any,
     ) -> bool {
-        let value_t = if let Some(value) = Self::take_t(value) {
-            value
+        if let Some(value_t) = take_option_t::<S::Component>(value) {
+            S::replace(self, entity, archetype, index, value_t).is_some()
         } else {
-            return false;
-        };
-        match self {
-            Storage::Dense(columns) => {
-                let col = vec_make_available(columns, archetype.index());
-                if let Some(entry) = col.get_mut(index) {
-                    *entry = value_t;
-                    true
-                } else {
-                    false
-                }
-            }
-            Storage::Sparse(map) => {
-                map.insert(entity, value_t);
-                true
-            }
+            false
         }
     }
 
     fn swap_remove_and_insert_to(
         &mut self,
-        _entity: Entity,
+        entity: Entity,
         remove_from_archetype: ArchetypeId,
         remove_from_index: usize,
         insert_to_archetype: ArchetypeId,
@@ -200,40 +269,11 @@ where
         if remove_from_archetype == insert_to_archetype {
             return None;
         }
-        if let Storage::Dense(columns) = self {
-            let from_len = columns.get(remove_from_archetype.index())?.len();
-            if remove_from_index < from_len {
-                vec_make_available(columns, insert_to_archetype.index());
-                if let Some((from_col, to_col)) = slice_get_mut2(
-                    columns,
-                    remove_from_archetype.index(),
-                    insert_to_archetype.index(),
-                ) {
-                    let to_index = to_col.len();
-                    to_col.push(from_col.swap_remove(remove_from_index));
-                    return Some(to_index);
-                }
-            }
+        if let Some(value) = S::swap_remove(self, entity, remove_from_archetype, remove_from_index)
+        {
+            Some(S::insert(self, entity, insert_to_archetype, value))
+        } else {
+            None
         }
-        None
-    }
-}
-
-#[inline]
-pub fn slice_get_mut2<T>(
-    slice: &mut [T],
-    index1: usize,
-    index2: usize,
-) -> Option<(&'_ mut T, &'_ mut T)> {
-    match index1.cmp(&index2) {
-        Ordering::Less => {
-            let (a, b) = slice.split_at_mut(index2);
-            Some((&mut a[index1], &mut b[0]))
-        }
-        Ordering::Greater => {
-            let (a, b) = slice.split_at_mut(index1);
-            Some((&mut b[0], &mut a[index2]))
-        }
-        Ordering::Equal => None,
     }
 }