+++ crates/ecs/CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this crate will be documented in this file.
 
 ## Unreleased
 
+ * Addes explicit Component trait and derive-macro
  * Split out Scheduling & Systems into own crate
  * Scheduling systems
  * `Query` for querying Components of a `World`